requirements.txt
numpy==1.26.3
scipy==0.16.0
python-dotenv==1.0.1
neo4j==5.27.0
streamlit==1.40.2
langchain==0.3.10
langchain_community==0.3.10
langchain-openai==0.2.11
pymongo==4.10.1
